pricetickerservice

+++ backend/Models/Product.cs
@@ -4,8 +4,8 @@ public class Product
     public int ProductID { get; set; }
     public string Naam { get; set; } = string.Empty;
     public string Beschrijving { get; set; } = string.Empty;
-    public float? StartPrijs { get; set; }
-    public float? MinPrijs {  get; set; }
+    public decimal? StartPrijs { get; set; }
+    public decimal? MinPrijs {  get; set; }
     public string? ImageUrl { get; set; }
 
     public float? Eindprijs { get; set; }
